+++ CHANGELOG.md
@@ -16,6 +16,12 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
 	https://changelog.md/
 -->
 
+## v0.6.0 (WIP)
+
+- Added status coloring on multiple statuses,
+  where `NotReady,SchedulingDisabled` would have red `NotReady`
+  and orange `SchedulingDisabled`. (#69)
+
 ## v0.5.1 (2023-11-11)
 
 - Fixes `--watch-kubeconfig` to reading a kubeconfig that has not yet been